don't crash when pausing a transfer and then removing it Patch by CCMAIL: Alexander Stillich <chills_@gmx.net> (thanks!)
svn path=/trunk/kdenetwork/kget/; revision=288594

+++ slave.cpp
@@ -129,8 +129,11 @@ void Slave::run()
             case KILL:
                 mDebug << " FETCHED COMMAND      KILL" << endl;
                 running = false;
-                copyjob->kill(true);
-                copyjob = 0L;
+                if (copyjob)
+                {
+                    copyjob->kill(true);
+                    copyjob = 0L;
+                 }
                 // no message posted
                 break;
